requirements.txt
# study-assistant-backend/requirements.txt

fastapi
uvicorn[standard]
motor
passlib[bcrypt]
pyjwt
python-dotenv
unstructured
sentence-transformers
pymilvus
stripe